refactor(motor-control): single source of truth for the MCP command numbers
The mcp266 CANopen driver mirrors the Basicmicro packet-serial command set into
its manufacturer object dictionary at 0x2000 + command number, but it hardcoded
those command numbers (24, 82, 200, 61-64, 32/33/35/36) as magic values that had
to be kept in lockstep with basicmicro's BasicmicroCommand enum by hand.

Move the BasicmicroCommand table into the shared motor_controller component
(basicmicro_commands.hpp) -- which both drivers already depend on, so neither
sibling depends on the other -- and derive the manufacturer object indices from
the named enum values via a command_object(BasicmicroCommand) overload. The
numbers now live in exactly one place and cannot drift.

No value changes: both host cores test-verify the same concrete addresses
(e.g. ReadMainBatteryVoltage -> 0x2018). basicmicro_core keeps its packet codecs
/ CRC / BasicmicroStatus (transport-specific, not shared).

Verified: basicmicro + mcp266 host tests pass; both examples build on IDF v6.0.1.
